+++ src/index.js
@@ -34,8 +34,8 @@ import facebookbotProvider from './providers/facebookbot.js';
 import brevoProvider from './providers/brevo.js';
 import getTermsProvider from './providers/get-terms.js';
 import labrikaProvider from './providers/labrika.js';
-// import mailgunProvider from './providers/mailgun.js';
-// import seobilityProvider from './providers/seobility.js'; // Unreliable
+import mailgunProvider from './providers/mailgun.js';
+import seobilityProvider from './providers/seobility.js';
 
 // Constants for IP address versions
 const IP_VERSION_V4 = 'ipv4';
@@ -81,8 +81,8 @@ const defaultProviders = [
   brevoProvider,
   getTermsProvider,
   labrikaProvider,
-  // mailgunProvider,
-  // seobilityProvider, // Unreliable
+  mailgunProvider,
+  seobilityProvider,
 ];
 
 const parsedAddresses = new LRUCache(MAX_PARSED_ADDRESSES);